Refactor: updated file structure to have more overview. Added helpers.py and schemas.py to limit lines of code in config_flow.py
