+++ custom_components/aio_energy_management/cheapest_hours/config_flow.py
@@ -1,50 +1,34 @@
-"""Cheapest hours config flow handlers and helpers."""
+"""Cheapest hours config flow handlers."""
 
 from __future__ import annotations
 
 import logging
 from typing import Any
 
-import voluptuous as vol
+from homeassistant.config_entries import ConfigFlowResult
 
-from homeassistant.config_entries import ConfigEntry, ConfigFlowResult
-from homeassistant.const import CONF_NAME
-from homeassistant.data_entry_flow import section
-from homeassistant.helpers import selector
-from homeassistant.core import HomeAssistant
-import homeassistant.helpers.config_validation as cv
-
-from ..const import (
+from custom_components.aio_energy_management.const import (
     CONF_ADD_FLEXIBLE,
     CONF_AREA,
-    CONF_CALENDAR,
     CONF_DATA_PROVIDER_TYPE,
-    CONF_END,
     CONF_END_HOURS_ENTITY,
     CONF_END_MINUTES_ENTITY,
-    CONF_ENTSOE_ENTITY,
     CONF_FAILSAFE_STARTING_HOUR,
-    CONF_FIRST_HOUR,
-    CONF_HOURS,
-    CONF_INVERSED,
-    CONF_LAST_HOUR,
+    CONF_FLEXIBLE_PRICE_LIMIT,
+    CONF_FLEXIBLE_PRICE_LIMIT_ENTITY,
     CONF_MAX_NUMBER_OF_SLOTS,
     CONF_MAX_NUMBER_OF_SLOTS_ENTITY,
-    CONF_MINUTES,
+    CONF_MIN_SEQ_SLOTS,
     CONF_MTU,
-    CONF_NORDPOOL_ENTITY,
-    CONF_NORDPOOL_OFFICIAL_CONFIG_ENTRY,
+    CONF_NAME,
+    CONF_NUMBER_OF_BLOCKS,
     CONF_NUMBER_OF_SLOTS,
     CONF_NUMBER_OF_SLOTS_ENTITY,
-    CONF_MIN_SEQ_SLOTS,
-    CONF_NUMBER_OF_BLOCKS,
     CONF_OFFSET,
     CONF_PRICE_LIMIT,
     CONF_PRICE_LIMIT_ENTITY,
     CONF_PRICE_MODIFICATIONS,
-    CONF_RETENTION_DAYS,
     CONF_SEQUENTIAL,
-    CONF_START,
     CONF_START_HOURS_ENTITY,
     CONF_START_MINUTES_ENTITY,
     CONF_TRIGGER_HOUR,
@@ -54,709 +38,34 @@ from ..const import (
     DATA_PROVIDER_NORDPOOL,
     DATA_PROVIDER_NORDPOOL_OFFICIAL,
     DATA_PROVIDER_STROMLIGNING,
-    CONF_STROMLIGNING_ENTITY,
-    CONF_STROMLIGNING_TOMORROW_ENTITY,
+)
+from .helpers import (
+    _coerce_mtu,
+    _normalize_optional_keys,
+    _process_offset_input,
+    _validate_advanced_integer_fields,
+    _validate_and_build_add_flexible,
+    _validate_and_clean_advanced_fields,
+    _validate_and_clean_number_of_slots,
+    _validate_and_clean_offset_fields,
+    _validate_basic_integer_fields,
+    _validate_offset_integer_fields,
+)
+from .schemas import (
+    _get_cheapest_hours_advanced_schema,
+    _get_cheapest_hours_basic_schema,
+    _get_data_provider_type_schema,
+    _get_entsoe_schema,
+    _get_nordpool_official_schema,
+    _get_nordpool_schema,
+    _get_offset_schema,
+    _get_stromligning_schema,
 )
 
 _LOGGER = logging.getLogger(__name__)
 
 CONF_ENTRY_TYPE = "entry_type"
 ENTRY_TYPE_CHEAPEST_HOURS = "cheapest_hours"
-
-# Flat UI field names for the flexible price limit. They are assembled into the
-# nested ``add_flexible`` dict (using CONF_PRICE_LIMIT / CONF_PRICE_LIMIT_ENTITY)
-# before the config entry is stored, so they cannot reuse the top-level
-# price_limit field names.
-CONF_FLEXIBLE_PRICE_LIMIT = "flexible_price_limit"
-CONF_FLEXIBLE_PRICE_LIMIT_ENTITY = "flexible_price_limit_entity"
-
-# Reusable selectors
-SEL_HOUR = selector.NumberSelector(
-    selector.NumberSelectorConfig(min=0, max=23, mode=selector.NumberSelectorMode.BOX)
-)
-SEL_FLOAT = selector.NumberSelector(
-    selector.NumberSelectorConfig(mode=selector.NumberSelectorMode.BOX, step="any")
-)
-SEL_INT = selector.NumberSelector(
-    selector.NumberSelectorConfig(min=1, mode=selector.NumberSelectorMode.BOX, step=1)
-)
-SEL_ENTITY = selector.EntitySelector(
-    selector.EntitySelectorConfig(domain=["sensor", "input_number"])
-)
-SEL_INT_0 = selector.NumberSelector(
-    selector.NumberSelectorConfig(min=0, mode=selector.NumberSelectorMode.BOX, step=1)
-)
-SEL_INT_ANY = selector.NumberSelector(
-    selector.NumberSelectorConfig(mode=selector.NumberSelectorMode.BOX, step=1)
-)
-SEL_SENSOR = selector.EntitySelector(selector.EntitySelectorConfig(domain="sensor"))
-SEL_BINARY_SENSOR = selector.EntitySelector(
-    selector.EntitySelectorConfig(domain="binary_sensor")
-)
-
-
-def _get_val(
-    data: dict, key: str, flex_key: str | None = None, default: Any = None
-) -> Any:
-    """Retrieve the suggested value from user_input, add_flexible, or the default."""
-    flex_data = data.get(CONF_ADD_FLEXIBLE) or {}
-    val = flex_data.get(flex_key) if flex_key else None
-    if val is None:
-        val = data.get(key)
-    return val if val is not None else default
-
-
-def _req(key: str, data: dict, default: Any = None) -> vol.Required:
-    """Create a `vol.Required` key with a `default` or `suggested_value`."""
-    val = data.get(key)
-    if default is not None:
-        return vol.Required(key, default=val if val is not None else default)
-    desc = {"suggested_value": val} if val is not None else {}
-    return vol.Required(key, description=desc)
-
-
-def _opt(
-    key: str, data: dict, flex_key: str | None = None, default: Any = None
-) -> vol.Optional:
-    """Create a `vol.Optional` key with the suggested value in a single line."""
-    val = _get_val(data, key, flex_key, default)
-    desc = {"suggested_value": val} if val is not None else {}
-    return vol.Optional(key, description=desc)
-
-
-def _get_data_provider_type_schema(default: str | None = None) -> vol.Schema:
-    """Get data provider type selection schema."""
-    req_key = (
-        vol.Required(CONF_DATA_PROVIDER_TYPE, default=default)
-        if default
-        else vol.Required(CONF_DATA_PROVIDER_TYPE)
-    )
-    return vol.Schema(
-        {
-            req_key: vol.In(
-                {
-                    DATA_PROVIDER_NORDPOOL: "Nord Pool",
-                    DATA_PROVIDER_NORDPOOL_OFFICIAL: "Nord Pool official",
-                    DATA_PROVIDER_ENTSOE: "Entso-E",
-                    DATA_PROVIDER_STROMLIGNING: "Strømligning",
-                }
-            ),
-        }
-    )
-
-
-def _mtu_selector() -> selector.SelectSelector:
-    """Return the MTU dropdown selector.
-
-    String option values are used (instead of ``vol.In([15, 60])``) because the
-    frontend only reliably pre-selects string-valued select options. The value
-    is coerced back to an int via ``_coerce_mtu`` in the step handler.
-    """
-    return selector.SelectSelector(
-        selector.SelectSelectorConfig(
-            options=["15", "60"],
-            mode=selector.SelectSelectorMode.DROPDOWN,
-        )
-    )
-
-
-def _mtu_default(user_input: dict[str, Any] | None) -> str:
-    """Return the MTU dropdown default as a string."""
-    if user_input and user_input.get(CONF_MTU) is not None:
-        return str(user_input[CONF_MTU])
-    return "60"
-
-
-def _coerce_mtu(user_input: dict[str, Any]) -> None:
-    """Coerce the MTU dropdown value (a string) back to an int in place."""
-    if user_input.get(CONF_MTU) is not None:
-        user_input[CONF_MTU] = int(user_input[CONF_MTU])
-
-
-def _get_nordpool_schema(user_input: dict[str, Any] | None = None) -> vol.Schema:
-    """Get Nord Pool entity selection schema."""
-    data = user_input or {}
-    return vol.Schema(
-        {
-            _req(CONF_NORDPOOL_ENTITY, data): SEL_SENSOR,
-            _opt(CONF_MTU, data, default=_mtu_default(user_input)): _mtu_selector(),
-        }
-    )
-
-
-def _get_nordpool_official_schema(
-    hass: HomeAssistant,
-    user_input: dict[str, Any] | None = None,
-) -> vol.Schema:
-    """Get Nord Pool official config entry ID schema."""
-    data = user_input or {}
-    existing_entries = hass.config_entries.async_entries("nordpool")
-    options = [{"value": e.entry_id, "label": e.title} for e in existing_entries]
-
-    return vol.Schema(
-        {
-            vol.Required(CONF_NORDPOOL_OFFICIAL_CONFIG_ENTRY): selector.SelectSelector(
-                selector.SelectSelectorConfig(
-                    options=options,
-                    mode=selector.SelectSelectorMode.DROPDOWN,
-                )
-            ),
-            _opt(CONF_AREA, data): cv.string,
-            _opt(CONF_MTU, data, default=_mtu_default(user_input)): _mtu_selector(),
-        }
-    )
-
-
-def _get_entsoe_schema(user_input: dict[str, Any] | None = None) -> vol.Schema:
-    """Get Entso-E entity selection schema."""
-    data = user_input or {}
-    return vol.Schema(
-        {
-            _req(CONF_ENTSOE_ENTITY, data): SEL_SENSOR,
-            _opt(CONF_MTU, data, default=_mtu_default(user_input)): _mtu_selector(),
-        }
-    )
-
-
-def _get_stromligning_schema(user_input: dict[str, Any] | None = None) -> vol.Schema:
-    """Get Strømligning entity selection schema."""
-    data = user_input or {}
-    return vol.Schema(
-        {
-            _req(CONF_STROMLIGNING_ENTITY, data): SEL_SENSOR,
-            _req(CONF_STROMLIGNING_TOMORROW_ENTITY, data): SEL_BINARY_SENSOR,
-            _opt(CONF_MTU, data, default=_mtu_default(user_input)): _mtu_selector(),
-        }
-    )
-
-
-def _get_cheapest_hours_basic_schema(
-    user_input: dict[str, Any] | None = None,
-) -> vol.Schema:
-    """Get basic cheapest hours configuration schema."""
-    data = user_input or {}
-    has_dynamic = bool(data.get(CONF_NUMBER_OF_SLOTS_ENTITY))
-
-    schema_dict = {
-        _req(CONF_NAME, data, default="Cheapest Hours"): cv.string,
-        _opt(CONF_NUMBER_OF_SLOTS, data, default=0): SEL_INT_0,
-        _req(CONF_FIRST_HOUR, data, default=0): SEL_HOUR,
-        _req(CONF_LAST_HOUR, data, default=23): SEL_HOUR,
-        _req(CONF_SEQUENTIAL, data, default=False): cv.boolean,
-        _req(CONF_CALENDAR, data, default=True): cv.boolean,
-        _req(CONF_INVERSED, data, default=False): cv.boolean,
-        vol.Required("dynamic_section"): section(
-            vol.Schema(
-                {
-                    _opt(CONF_NUMBER_OF_SLOTS_ENTITY, data): SEL_ENTITY,
-                }
-            ),
-            {"collapsed": not has_dynamic},
-        ),
-    }
-
-    return vol.Schema(schema_dict)
-
-
-def _get_cheapest_hours_advanced_schema(
-    user_input: dict[str, Any] | None = None,
-    sequential: bool = False,
-) -> vol.Schema:
-    """Get advanced cheapest hours configuration schema."""
-    data = user_input or {}
-
-    # 1. Main settings
-    schema_dict = {
-        _opt(CONF_FAILSAFE_STARTING_HOUR, data): SEL_HOUR,
-        _opt(CONF_TRIGGER_HOUR, data): SEL_HOUR,
-        _opt(CONF_PRICE_LIMIT, data): SEL_FLOAT,
-    }
-
-    # 2. Non-sequential fields
-    if not sequential:
-        schema_dict.update(
-            {
-                _opt(
-                    CONF_FLEXIBLE_PRICE_LIMIT, data, flex_key=CONF_PRICE_LIMIT
-                ): SEL_FLOAT,
-                _opt(
-                    CONF_MAX_NUMBER_OF_SLOTS, data, flex_key=CONF_MAX_NUMBER_OF_SLOTS
-                ): SEL_INT,
-                _opt(CONF_MIN_SEQ_SLOTS, data): SEL_INT,
-                _opt(CONF_NUMBER_OF_BLOCKS, data): SEL_INT,
-            }
-        )
-
-    schema_dict.update(
-        {
-            _opt(CONF_RETENTION_DAYS, data, default=1): selector.NumberSelector(
-                selector.NumberSelectorConfig(
-                    min=1, max=365, mode=selector.NumberSelectorMode.BOX, step=1
-                )
-            ),
-            _opt(CONF_PRICE_MODIFICATIONS, data): selector.TemplateSelector(),
-        }
-    )
-
-    # 3. Dynamic entity section
-    dynamic_dict = {
-        _opt(CONF_TRIGGER_HOUR_ENTITY, data): SEL_ENTITY,
-        _opt(CONF_PRICE_LIMIT_ENTITY, data): SEL_ENTITY,
-    }
-
-    if not sequential:
-        dynamic_dict.update(
-            {
-                _opt(
-                    CONF_FLEXIBLE_PRICE_LIMIT_ENTITY,
-                    data,
-                    flex_key=CONF_PRICE_LIMIT_ENTITY,
-                ): SEL_ENTITY,
-                _opt(
-                    CONF_MAX_NUMBER_OF_SLOTS_ENTITY,
-                    data,
-                    flex_key=CONF_MAX_NUMBER_OF_SLOTS_ENTITY,
-                ): SEL_ENTITY,
-            }
-        )
-
-    # 4. Determine whether the dynamic section should be expanded.
-    has_dynamic = any(
-        _get_val(data, key, flex)
-        for key, flex in [
-            (CONF_TRIGGER_HOUR_ENTITY, None),
-            (CONF_PRICE_LIMIT_ENTITY, None),
-            (CONF_MAX_NUMBER_OF_SLOTS_ENTITY, CONF_MAX_NUMBER_OF_SLOTS_ENTITY),
-            (CONF_FLEXIBLE_PRICE_LIMIT_ENTITY, CONF_PRICE_LIMIT_ENTITY),
-        ]
-    )
-
-    schema_dict[vol.Required("dynamic_section")] = section(
-        vol.Schema(dynamic_dict),
-        {"collapsed": not has_dynamic},
-    )
-
-    return vol.Schema(schema_dict)
-
-
-def _process_offset_input(
-    user_input: dict[str, Any],
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    """Process offset input and build offset structure.
-
-    Returns:
-        Tuple of (offset_dict, entity_dict) where offset_dict contains static values
-        and entity_dict contains entity references to be stored at root level.
-    """
-    offset = {}
-    entities = {}
-
-    if any(
-        user_input.get(f"{CONF_START}_{key}") is not None
-        for key in [CONF_HOURS, CONF_MINUTES]
-    ):
-        start_offset = {}
-        if user_input.get(f"{CONF_START}_{CONF_HOURS}") is not None:
-            start_offset[CONF_HOURS] = user_input[f"{CONF_START}_{CONF_HOURS}"]
-        if user_input.get(f"{CONF_START}_{CONF_MINUTES}") is not None:
-            start_offset[CONF_MINUTES] = user_input[f"{CONF_START}_{CONF_MINUTES}"]
-        if start_offset:
-            offset[CONF_START] = start_offset
-
-    if user_input.get(CONF_START_HOURS_ENTITY):
-        entities[CONF_START_HOURS_ENTITY] = user_input[CONF_START_HOURS_ENTITY]
-    if user_input.get(CONF_START_MINUTES_ENTITY):
-        entities[CONF_START_MINUTES_ENTITY] = user_input[CONF_START_MINUTES_ENTITY]
-
-    if any(
-        user_input.get(f"{CONF_END}_{key}") is not None
-        for key in [CONF_HOURS, CONF_MINUTES]
-    ):
-        end_offset = {}
-        if user_input.get(f"{CONF_END}_{CONF_HOURS}") is not None:
-            end_offset[CONF_HOURS] = user_input[f"{CONF_END}_{CONF_HOURS}"]
-        if user_input.get(f"{CONF_END}_{CONF_MINUTES}") is not None:
-            end_offset[CONF_MINUTES] = user_input[f"{CONF_END}_{CONF_MINUTES}"]
-        if end_offset:
-            offset[CONF_END] = end_offset
-
-    if user_input.get(CONF_END_HOURS_ENTITY):
-        entities[CONF_END_HOURS_ENTITY] = user_input[CONF_END_HOURS_ENTITY]
-    if user_input.get(CONF_END_MINUTES_ENTITY):
-        entities[CONF_END_MINUTES_ENTITY] = user_input[CONF_END_MINUTES_ENTITY]
-
-    return offset, entities
-
-
-def _get_offset_schema(offset_data: dict[str, Any]) -> vol.Schema:
-    """Get offset configuration schema."""
-    start = offset_data.get(CONF_START, {})
-    end = offset_data.get(CONF_END, {})
-
-    # Combine into a flat dictionary so that `_opt` works directly everywhere.
-    data = {
-        f"{CONF_START}_{CONF_HOURS}": start.get(CONF_HOURS),
-        f"{CONF_START}_{CONF_MINUTES}": start.get(CONF_MINUTES),
-        f"{CONF_END}_{CONF_HOURS}": end.get(CONF_HOURS),
-        f"{CONF_END}_{CONF_MINUTES}": end.get(CONF_MINUTES),
-        **offset_data,
-    }
-
-    # 1. Static offset fields (hour / minute)
-    schema_dict = {
-        _opt(f"{CONF_START}_{CONF_HOURS}", data): SEL_INT_ANY,
-        _opt(f"{CONF_START}_{CONF_MINUTES}", data): SEL_INT_ANY,
-        _opt(f"{CONF_END}_{CONF_HOURS}", data): SEL_INT_ANY,
-        _opt(f"{CONF_END}_{CONF_MINUTES}", data): SEL_INT_ANY,
-    }
-
-    # 2. Dynamic entities
-    dynamic_dict = {
-        _opt(CONF_START_HOURS_ENTITY, data): SEL_ENTITY,
-        _opt(CONF_START_MINUTES_ENTITY, data): SEL_ENTITY,
-        _opt(CONF_END_HOURS_ENTITY, data): SEL_ENTITY,
-        _opt(CONF_END_MINUTES_ENTITY, data): SEL_ENTITY,
-    }
-
-    # 3. Check whether the dynamic section needs to be expanded.
-    has_dynamic = any(
-        data.get(key)
-        for key in (
-            CONF_START_HOURS_ENTITY,
-            CONF_START_MINUTES_ENTITY,
-            CONF_END_HOURS_ENTITY,
-            CONF_END_MINUTES_ENTITY,
-        )
-    )
-
-    schema_dict[vol.Required("dynamic_section")] = section(
-        vol.Schema(dynamic_dict),
-        {"collapsed": not has_dynamic},
-    )
-
-    return vol.Schema(schema_dict)
-
-
-def _validate_and_clean_static_or_entity(
-    user_input: dict[str, Any],
-    static_key: str,
-    entity_key: str,
-    field_name: str,
-    allow_both_empty: bool = False,
-) -> dict[str, str]:
-    """Validate and clean configuration where either static value or entity can be used.
-
-    Args:
-        user_input: The user input dictionary (will be modified to remove unused fields)
-        static_key: Key for the static value field
-        entity_key: Key for the entity field
-        field_name: Human-readable field name for error messages
-        allow_both_empty: If True, allows both fields to be empty
-
-    Returns:
-        Dict of errors (empty if validation passes).
-    """
-    errors: dict[str, str] = {}
-
-    static_value = user_input.get(static_key)
-    has_static = static_value is not None and (
-        (isinstance(static_value, (int, float)) and static_value != 0)
-        or (isinstance(static_value, str) and static_value.strip())
-    )
-
-    has_entity = bool(user_input.get(entity_key))
-
-    if has_static and has_entity:
-        errors["base"] = f"both_{field_name}_configured"
-    elif not has_static and not has_entity and not allow_both_empty:
-        errors["base"] = f"no_{field_name}_configured"
-    else:
-        if has_entity and static_key in user_input:
-            user_input.pop(static_key, None)
-        elif has_static and entity_key in user_input:
-            user_input.pop(entity_key, None)
-        elif not has_static and not has_entity and allow_both_empty:
-            user_input.pop(static_key, None)
-            user_input.pop(entity_key, None)
-
-    return errors
-
-
-def _validate_and_clean_number_of_slots(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate and clean number of slots configuration.
-
-    Removes unused fields and returns a dict of errors (empty if validation passes).
-    """
-    return _validate_and_clean_static_or_entity(
-        user_input,
-        CONF_NUMBER_OF_SLOTS,
-        CONF_NUMBER_OF_SLOTS_ENTITY,
-        "slots",
-        allow_both_empty=False,
-    )
-
-
-def _validate_and_clean_advanced_fields(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate and clean advanced configuration fields.
-
-    Validates trigger_hour and price_limit (both optional, can use static or entity).
-    """
-    errors: dict[str, str] = {}
-
-    trigger_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        CONF_TRIGGER_HOUR,
-        CONF_TRIGGER_HOUR_ENTITY,
-        "trigger_hour",
-        allow_both_empty=True,
-    )
-    if trigger_errors:
-        errors.update(trigger_errors)
-
-    price_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        CONF_PRICE_LIMIT,
-        CONF_PRICE_LIMIT_ENTITY,
-        "price_limit",
-        allow_both_empty=True,
-    )
-    if price_errors:
-        errors.update(price_errors)
-
-    return errors
-
-
-def _validate_and_build_add_flexible(
-    user_input: dict[str, Any],
-    mtu: int,
-) -> dict[str, str]:
-    """Validate flexible slot fields and assemble them into a nested dict.
-
-    The flat ``max_number_of_slots`` / ``flexible_price_limit`` UI fields (and
-    their entity variants) are validated and, on success, collapsed into a
-    nested ``add_flexible`` dict stored on ``user_input``. Both the max slots
-    and the price limit must be provided together, since one without the other
-    has no effect.
-    """
-    errors: dict[str, str] = {}
-
-    max_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        CONF_MAX_NUMBER_OF_SLOTS,
-        CONF_MAX_NUMBER_OF_SLOTS_ENTITY,
-        "max_number_of_slots",
-        allow_both_empty=True,
-    )
-    errors.update(max_errors)
-
-    price_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        CONF_FLEXIBLE_PRICE_LIMIT,
-        CONF_FLEXIBLE_PRICE_LIMIT_ENTITY,
-        "flexible_price_limit",
-        allow_both_empty=True,
-    )
-    errors.update(price_errors)
-
-    if errors:
-        return errors
-
-    has_max = (
-        CONF_MAX_NUMBER_OF_SLOTS in user_input
-        or CONF_MAX_NUMBER_OF_SLOTS_ENTITY in user_input
-    )
-    has_price = (
-        CONF_FLEXIBLE_PRICE_LIMIT in user_input
-        or CONF_FLEXIBLE_PRICE_LIMIT_ENTITY in user_input
-    )
-
-    if has_max != has_price:
-        errors["base"] = "add_flexible_incomplete"
-        return errors
-
-    max_slots = user_input.get(CONF_MAX_NUMBER_OF_SLOTS)
-    if max_slots is not None:
-        cap = 96 if mtu == 15 else 24
-        if max_slots < 1 or max_slots > cap:
-            errors[CONF_MAX_NUMBER_OF_SLOTS] = "max_number_of_slots_out_of_range"
-            return errors
-
-    # The flexible price limit only extends the base slots with cheaper ones, so
-    # a static flexible limit must stay below the regular (static) price limit.
-    flexible_price_limit = user_input.get(CONF_FLEXIBLE_PRICE_LIMIT)
-    price_limit = user_input.get(CONF_PRICE_LIMIT)
-    if (
-        flexible_price_limit is not None
-        and price_limit is not None
-        and flexible_price_limit >= price_limit
-    ):
-        errors[CONF_FLEXIBLE_PRICE_LIMIT] = "flexible_price_limit_not_below_price_limit"
-        return errors
-
-    add_flexible: dict[str, Any] = {}
-    if CONF_MAX_NUMBER_OF_SLOTS in user_input:
-        add_flexible[CONF_MAX_NUMBER_OF_SLOTS] = user_input.pop(
-            CONF_MAX_NUMBER_OF_SLOTS
-        )
-    if CONF_MAX_NUMBER_OF_SLOTS_ENTITY in user_input:
-        add_flexible[CONF_MAX_NUMBER_OF_SLOTS_ENTITY] = user_input.pop(
-            CONF_MAX_NUMBER_OF_SLOTS_ENTITY
-        )
-    if CONF_FLEXIBLE_PRICE_LIMIT in user_input:
-        add_flexible[CONF_PRICE_LIMIT] = user_input.pop(CONF_FLEXIBLE_PRICE_LIMIT)
-    if CONF_FLEXIBLE_PRICE_LIMIT_ENTITY in user_input:
-        add_flexible[CONF_PRICE_LIMIT_ENTITY] = user_input.pop(
-            CONF_FLEXIBLE_PRICE_LIMIT_ENTITY
-        )
-
-    if add_flexible:
-        user_input[CONF_ADD_FLEXIBLE] = add_flexible
-    else:
-        user_input.pop(CONF_ADD_FLEXIBLE, None)
-
-    return errors
-
-
-def _validate_and_clean_offset_fields(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate and clean offset configuration fields.
-
-    Validates start/end hours and minutes (all optional, can use static or entity).
-    """
-    errors: dict[str, str] = {}
-
-    start_hours_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        f"{CONF_START}_{CONF_HOURS}",
-        CONF_START_HOURS_ENTITY,
-        "start_hours",
-        allow_both_empty=True,
-    )
-    if start_hours_errors:
-        errors.update(start_hours_errors)
-
-    start_minutes_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        f"{CONF_START}_{CONF_MINUTES}",
-        CONF_START_MINUTES_ENTITY,
-        "start_minutes",
-        allow_both_empty=True,
-    )
-    if start_minutes_errors:
-        errors.update(start_minutes_errors)
-
-    end_hours_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        f"{CONF_END}_{CONF_HOURS}",
-        CONF_END_HOURS_ENTITY,
-        "end_hours",
-        allow_both_empty=True,
-    )
-    if end_hours_errors:
-        errors.update(end_hours_errors)
-
-    end_minutes_errors = _validate_and_clean_static_or_entity(
-        user_input,
-        f"{CONF_END}_{CONF_MINUTES}",
-        CONF_END_MINUTES_ENTITY,
-        "end_minutes",
-        allow_both_empty=True,
-    )
-    if end_minutes_errors:
-        errors.update(end_minutes_errors)
-
-    return errors
-
-
-# ---------------------------------------------------------------------------
-# Integer validation helpers
-# These perform custom validation instead of using vol.Range so that integer
-# fields keep their text-box appearance (vol.Range would turn them into sliders).
-# ---------------------------------------------------------------------------
-
-
-def _validate_basic_integer_fields(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate integer fields on the basic cheapest hours step.
-
-    Checks:
-    - first_hour: 0-23
-    - last_hour: 0-23
-    - number_of_slots: >= 0
-
-    Note: last_hour may be less than first_hour to support overnight windows
-    (e.g. first_hour=22, last_hour=6).
-    """
-    errors: dict[str, str] = {}
-
-    first_hour = user_input.get(CONF_FIRST_HOUR)
-    last_hour = user_input.get(CONF_LAST_HOUR)
-    number_of_slots = user_input.get(CONF_NUMBER_OF_SLOTS)
-
-    if first_hour is not None and not (0 <= first_hour <= 23):
-        errors[CONF_FIRST_HOUR] = "first_hour_out_of_range"
-
-    if last_hour is not None and not (0 <= last_hour <= 23):
-        errors[CONF_LAST_HOUR] = "last_hour_out_of_range"
-
-    if number_of_slots is not None and number_of_slots < 0:
-        errors[CONF_NUMBER_OF_SLOTS] = "number_of_slots_negative"
-
-    return errors
-
-
-def _validate_advanced_integer_fields(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate optional integer fields on the advanced cheapest hours step.
-
-    Checks:
-    - failsafe_starting_hour: 0-23 (optional)
-    - trigger_hour: 0-23 (optional)
-    """
-    errors: dict[str, str] = {}
-
-    failsafe = user_input.get(CONF_FAILSAFE_STARTING_HOUR)
-    if failsafe is not None and not (0 <= failsafe <= 23):
-        errors[CONF_FAILSAFE_STARTING_HOUR] = "failsafe_starting_hour_out_of_range"
-
-    trigger_hour = user_input.get(CONF_TRIGGER_HOUR)
-    if trigger_hour is not None and not (0 <= trigger_hour <= 23):
-        errors[CONF_TRIGGER_HOUR] = "trigger_hour_out_of_range"
-
-    min_seq_slots = user_input.get(CONF_MIN_SEQ_SLOTS)
-    if min_seq_slots is not None and min_seq_slots < 1:
-        errors[CONF_MIN_SEQ_SLOTS] = "min_seq_slots_out_of_range"
-
-    number_of_blocks = user_input.get(CONF_NUMBER_OF_BLOCKS)
-    if number_of_blocks is not None and number_of_blocks < 1:
-        errors[CONF_NUMBER_OF_BLOCKS] = "number_of_blocks_out_of_range"
-
-    return errors
-
-
-def _validate_offset_integer_fields(user_input: dict[str, Any]) -> dict[str, str]:
-    """Validate optional integer fields on the offset step.
-
-    Checks:
-    - start_minutes: 0-59 (optional)
-    - end_minutes: 0-59 (optional)
-    Hours fields are unrestricted integers (they can exceed 23 for multi-hour offsets).
-    """
-    errors: dict[str, str] = {}
-
-    start_minutes = user_input.get(f"{CONF_START}_{CONF_MINUTES}")
-    if start_minutes is not None and not (-59 <= start_minutes <= 59):
-        errors[f"{CONF_START}_{CONF_MINUTES}"] = "start_minutes_out_of_range"
-
-    end_minutes = user_input.get(f"{CONF_END}_{CONF_MINUTES}")
-    if end_minutes is not None and not (-59 <= end_minutes <= 59):
-        errors[f"{CONF_END}_{CONF_MINUTES}"] = "end_minutes_out_of_range"
-
-    return errors
-
-
-def _normalize_optional_keys(user_input: dict[str, Any], keys: list[str]) -> None:
-    """Ensure every optional key is present (as None) so clearing it in the
-    UI actually overwrites the stored value instead of being merged away."""
-    for key in keys:
-        user_input.setdefault(key, None)
 
 
 class CheapestHoursConfigFlowMixin:
@@ -826,12 +135,10 @@ class CheapestHoursConfigFlowMixin:
         if user_input is not None:
             _coerce_mtu(user_input)
 
-            # In Options Flow: close and save entry
             if hasattr(self, "_config_entry"):
                 _normalize_optional_keys(user_input, [CONF_AREA])
                 return self._save_options_entry(user_input)
 
-            # In Config Flow: go to advanced
             self._config_data.update(user_input)
             return await self.async_step_cheapest_hours_basic()
 
@@ -898,13 +205,11 @@ class CheapestHoursConfigFlowMixin:
             errors.update(slot_errors)
 
             if not errors:
-                # In Options Flow: close and save entry
                 if hasattr(self, "_config_entry"):
                     _normalize_optional_keys(
                         user_input,
                         [CONF_NUMBER_OF_SLOTS, CONF_NUMBER_OF_SLOTS_ENTITY],
                     )
-                    # Force optional sequential fields to empty when flowing from non-sequential to sequential
                     if user_input.get(CONF_SEQUENTIAL) is True:
                         user_input[CONF_MIN_SEQ_SLOTS] = None
                         user_input[CONF_NUMBER_OF_BLOCKS] = None
@@ -915,7 +220,6 @@ class CheapestHoursConfigFlowMixin:
                         user_input[CONF_FLEXIBLE_PRICE_LIMIT_ENTITY] = None
                     return self._save_options_entry(user_input)
 
-                # In Config Flow: close and save entry
                 self._config_data.update(user_input)
 
                 unique_id = self._config_data[CONF_NAME].lower().replace(" ", "_")
